requirements.txt
Flask==2.0.1
requests==2.26.0
Werkzeug==2.0.1
python-dotenv==0.19.0
gunicorn==20.1.0
huggingface_hub==0.14.1